[REFACTOR]: fix conditional export issue

+++ script/build.ts
@@ -10,7 +10,7 @@ require("./lint");
 
 function cleanup() {
     logger.task("Cleanup Old Builds");
-    if (fs.statSync(path.join(__dirname, "../dist")).isDirectory()) {
+    if (fs.existsSync(path.join(__dirname, "../dist")) && fs.statSync(path.join(__dirname, "../dist")).isDirectory()) {
         fs.rmSync(path.join(__dirname, "../dist"), { recursive: true });
     }
 }
@@ -20,19 +20,20 @@ function build() {
     spawn("tsc", ["--build", path.join(__dirname, "../tsconfig.json")], "Build Failed.");
 }
 
-function copyCJStoLib() {
-    logger.task("Copying dist files to lib");
-    const dist = path.join(__dirname, "../dist");
-    fs.copySync(path.join(dist, "./cjs"), path.join(dist, "./lib"), { recursive: true });
-}
-
 function copyPackageJSON() {
     logger.task("Copying other asset to dist");
     fs.copyFileSync(path.join(__dirname, "../package.json"), path.join(__dirname, "../dist/package.json"));
     fs.copyFileSync(path.join(__dirname, "../LICENSE"), path.join(__dirname, "../dist/LICENSE"));
 }
 
+function copyDTS() {
+    logger.task("Copying d.ts to dist/");
+    fs.copyFileSync(path.join(__dirname, "./template/react.d.ts"), path.join(__dirname, "../dist/react.d.ts"));
+    fs.copyFileSync(path.join(__dirname, "./template/core.d.ts"), path.join(__dirname, "../dist/core.d.ts"));
+    fs.copyFileSync(path.join(__dirname, "./template/browser.d.ts"), path.join(__dirname, "../dist/browser.d.ts"));
+}
+
 cleanup();
 build();
-copyCJStoLib();
+copyDTS();
 copyPackageJSON();
